requirements.txt
pyodbc>=5.0.1
PyQt5>=5.15.10
azure-identity>=1.15.0
pyinstaller>=6.3.0
